# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Platform API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/exercises": {
            "get": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Get all exercises",
                "operationId": "21bd87ce16a1f663c1699749f963152d",
                "responses": {
                    "200": {
                        "description": "List of exercises"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Create a new exercise",
                "operationId": "1eac3da7158aade43d25b9ee98b2fb6f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "content",
                                    "lesson_id",
                                    "subject",
                                    "level",
                                    "points"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "lesson_id": {
                                        "type": "integer"
                                    },
                                    "subject": {
                                        "type": "string"
                                    },
                                    "level": {
                                        "type": "string"
                                    },
                                    "points": {
                                        "type": "integer"
                                    },
                                    "file_url": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exercise created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exercises/{id}": {
            "get": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Get exercise details",
                "operationId": "f3924e81949e4458f8638b614ca1940b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exercise details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exercises/search/by": {
            "get": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Search exercises",
                "operationId": "96edf891762ea8560ca9564b5474f2fb",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "subject",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exercises/assign-points": {
            "post": {
                "tags": [
                    "Exercises"
                ],
                "summary": "Assign points to student",
                "operationId": "0d62a25405953db793fdbc0c95ad70fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "student_id",
                                    "exercise_id",
                                    "points_earned"
                                ],
                                "properties": {
                                    "student_id": {
                                        "type": "integer"
                                    },
                                    "exercise_id": {
                                        "type": "integer"
                                    },
                                    "points_earned": {
                                        "type": "integer"
                                    },
                                    "teacher_notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Points assigned successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}